backend/api/requirements.txt
google-cloud-documentai
langchain
openai
pdf2image
opencv-python
python-dotenv
